requirements.txt
absl-py==0.9.0
aioredis==1.3.1
asgiref==3.2.7
astroid==2.3.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.2
Automat==20.2.0
autopep8==1.5.2
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
confusable-homoglyphs==3.2.0
constantly==15.1.0
crayons==0.3.0
cryptography==2.9
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-registration==3.1
django-rest-auth==0.9.5
django-versatileimagefield==2.0
django-webpack-loader==0.7.0
djangorestframework==3.11.0
gast==0.3.3
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
isort==4.3.21
joblib==0.15.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
msgpack==0.6.2
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
Pillow==7.1.1
protobuf==3.12.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
python-magic-bin==0.4.14
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.0
scipy==1.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.0.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
urllib3==1.25.8
webdriver-manager==2.4.0
Werkzeug==1.0.1
wrapt==1.11.2
zipp==3.1.0
zope.interface==5.1.0
